stating refactor to worked based on json rather that static class -> source -> site / api && factory producing said sites and apis configurations
